refactor: restructure store
