--- src/actions/index.js
@@ -1,4 +1,0 @@
-import { SIGN_IN, SIGN_OUT } from './types';
-
-export const signIn = () => ({ type: SIGN_IN });
-export const signOut = () => ({ type: SIGN_OUT });
