--- src/actions/types.js
@@ -1,2 +1,0 @@
-export const SIGN_IN = 'SIGN_IN';
-export const SIGN_OUT = 'SIGN_OUT';
