+++ src/index.tsx
@@ -4,7 +4,7 @@ import './index.css';
 import App from './App';
 import * as serviceWorker from './serviceWorker';
 import { Provider } from 'react-redux';
-import store from './app/store';
+import store from './redux/store';
 
 ReactDOM.render(
 	<React.StrictMode>
