+++ src/redux/store.ts
@@ -1,7 +1,8 @@
 import { configureStore } from '@reduxjs/toolkit';
-import AuthReducer from '../features/AuthSlice';
-import EventsListReducer from '../features/EventsListSlice';
-import ChannelListReducer from '../features/ChannelsSlice';
+import AuthReducer from './slices/AuthSlice';
+import EventsListReducer from './slices/EventsListSlice';
+import ChannelListReducer from './slices/ChannelsSlice';
+
 export default configureStore({
 	reducer: {
 		auth: AuthReducer,
